Extracted EventSource and Observable into own files
