+++ collector/src/main/java/com/bitmovin/analytics/Observable.kt
@@ -1,0 +1,6 @@
+package com.bitmovin.analytics
+
+interface Observable<TListener> {
+    fun subscribe(listener: TListener)
+    fun unsubscribe(listener: TListener)
+}
